migrate proto conversion module
