+++ src-tauri/client-proto/Cargo.toml
@@ -19,3 +19,6 @@ tonic.workspace = true
 tonic-prost.workspace = true
 
 defguard_wireguard_rs.workspace = true
+
+[dev-dependencies]
+x25519-dalek = { version = "2", features = ["getrandom", "static_secrets"] }
